story JS class 분리

+++ client/controller/story/story.controller.js
@@ -1,6 +1,7 @@
 import * as nodeFetch from "../../fetch/nodeFetch.js";
 import fs from "fs";
 import { config } from "../../config.js";
+import fetch from "node-fetch";
 
 export async function rederBoarePage(req, res, next) {
   res.status(200).render("story/story.board.ejs");
@@ -36,28 +37,17 @@ export async function postStory(req, res, next) {
   const token = req.cookies["token"];
   const bodyFilteredNotEmpty = filterNotEmptyString(req.body);
   const postData = { ...bodyFilteredNotEmpty, imgnames };
-  const responseFromPostStoryAPI = await nodeFetch.fetchPostApiWithToken(
-    postStoryURL,
-    postData,
-    token
-  );
+  const responseFromPostStoryAPI = await nodeFetch.fetchPostApiWithToken(postStoryURL, postData, token);
   if (responseFromPostStoryAPI.status === 201) {
     const responseParsed = await responseFromPostStoryAPI.json();
     const storyId = responseParsed.storyId;
     const isexist = fs.existsSync(`./uploads/story/story_${storyId}`);
     if (isexist) {
-      return res
-        .status(500)
-        .json({ msg: "Image directory of this id already exists" });
+      return res.status(500).json({ msg: "Image directory of this id already exists" });
     } else {
-      fs.mkdirSync(`./uploads/story/story_${storyId}`, (err) =>
-        res.status(500).json(err)
-      );
+      fs.mkdirSync(`./uploads/story/story_${storyId}`, (err) => res.status(500).json(err));
       imgnames.forEach((imgname) => {
-        fs.renameSync(
-          `./uploads/story/storyImg_temp/${imgname}`,
-          `./uploads/story/story_${storyId}/${imgname}`
-        );
+        fs.renameSync(`./uploads/story/storyImg_temp/${imgname}`, `./uploads/story/story_${storyId}/${imgname}`);
       });
       return res.redirect("/story/detail/" + storyId);
     }
@@ -71,10 +61,7 @@ export async function updateStory(req, res) {
   const id = req.params.id;
   const imgnames = req.files.map((img) => img.filename);
   imgnames.forEach((imgname) => {
-    fs.renameSync(
-      `./uploads/story/storyImg_temp/${imgname}`,
-      `./uploads/story/story_${id}/${imgname}`
-    );
+    fs.renameSync(`./uploads/story/storyImg_temp/${imgname}`, `./uploads/story/story_${id}/${imgname}`);
   });
   const putStoryURL = config.backendURL + "/story/" + id;
   const token = req.cookies["token"];
@@ -87,17 +74,11 @@ export async function updateStory(req, res) {
     deleteImgnames.push(deleteImgnamesInReqBody);
   }
   deleteImgnames.forEach((deleteImgname) => {
-    fs.unlink(`./uploads/story/story_${id}/${deleteImgname}`, (err) =>
-      console.error(err)
-    );
+    fs.unlink(`./uploads/story/story_${id}/${deleteImgname}`, (err) => console.error(err));
   });
   const bodyFilteredNotEmpty = filterNotEmptyString(req.body);
   const postData = { ...bodyFilteredNotEmpty, imgnames, deleteImgnames };
-  const responseFromPutStoryAPI = await nodeFetch.fetchPutApiWithToken(
-    putStoryURL,
-    postData,
-    token
-  );
+  const responseFromPutStoryAPI = await nodeFetch.fetchPutApiWithToken(putStoryURL, postData, token);
   const { storyId } = await responseFromPutStoryAPI.json();
   if (responseFromPutStoryAPI.status === 200) {
     return res.redirect("/story/detail/" + storyId);
@@ -113,8 +94,15 @@ export async function updateStory(req, res) {
 }
 
 export async function deleteStory(req, res, next) {
-  fs.rmSync(`./uploads/story/story_${req.params.id}`, {
-    recursive: true,
-    force: true,
-  });
+  try {
+    const { IsAuthor } = await fetch(config.backendURL + "/shop/author/" + req.params.id, { method: "GET" });
+    if (IsAuthor) {
+      fs.rmSync(`./uploads/story/story_${req.params.id}`, {
+        recursive: true,
+        force: true,
+      });
+    }
+  } catch (error) {
+    console.error(error);
+  }
 }
